agents/1_setup/requirements.txt
azure-ai-projects 
azure-identity 
python-dotenv
jsonref